requirements.txt
pylint
pyyaml
PyGithub
